+++ frontend/webpack.config.js
@@ -7,8 +7,6 @@ const TerserPlugin = require('terser-webpack-plugin');
 const OptimizeCSSAssetsPlugin = require("optimize-css-assets-webpack-plugin");
 const VueLoaderPlugin = require('vue-loader/lib/plugin');
 const LicenseWebpackPlugin = require('license-webpack-plugin').LicenseWebpackPlugin;
-const GoogleFontsPlugin = require("@beyonk/google-fonts-webpack-plugin");
-const CreateFileWebpack = require('create-file-webpack');
 
 module.exports = (env, argv) => {
     const devMode = argv.mode !== 'production';
@@ -57,7 +55,6 @@ module.exports = (env, argv) => {
                     'sass-loader',
                 ]
             }, {
-                // for font awesome fonts
                 test: /\.(woff|woff2|eot|ttf|otf|svg)$/,
                 use: [{
                     loader: 'file-loader',
@@ -101,6 +98,11 @@ module.exports = (env, argv) => {
                 filename: '[name].[contenthash].css'
             }),
             new VueLoaderPlugin(),
+            new CleanWebpackPlugin({
+                cleanOnceBeforeBuildPatterns: ['**/*', '../fonts/*'],
+                dangerouslyAllowCleanPatternsOutsideProject: true,
+                dry: false,
+            }),
         ],
         optimization: {
             runtimeChunk: 'single',
@@ -122,46 +124,8 @@ module.exports = (env, argv) => {
             ignored: [/node_modules/, /neucore-js-client/]
         }
     };
-
-    // do not remove and rebuild google fonts during watch mode
-    const cleanPattern = ['**/*', '../fonts/*'];
-    if (argv.watch) {
-        cleanPattern.push('!fonts.css');
-        cleanPattern.push('!font/**');
-    }
-    config.plugins.push(new CleanWebpackPlugin({
-        cleanOnceBeforeBuildPatterns: cleanPattern,
-        dangerouslyAllowCleanPatternsOutsideProject: true,
-        dry: false,
-    }));
-    if (! argv.watch) {
-        config.plugins.push(new GoogleFontsPlugin({
-            fonts: [
-                { family: "Source Sans Pro", variants: ["300", "400", "700", "400italic"] }, // cosmo, lumen
-                { family: "Roboto", variants: ["300", "400", "500", "700"] }, // cyborg, materia, sandstone
-                { family: "Lato", variants: ["300", "400", "700", "400italic"] }, // darkly, flatly, superhero
-                { family: "News Cycle", variants: ["400", "700"] }, // journal
-                { family: "Nunito Sans", variants: ["400", "600"] }, // lux
-                { family: "Montserrat" }, // minty
-                { // simplex, spacelab, yeti
-                    family: "Open Sans",
-                    variants: ["300", "400", "700", "300italic", "400italic", "700italic"]
-                },
-                { family: "Neucha" }, // sketchy
-                { family: "Cabin Sketch" }, // sketchy
-                { family: "Source Sans Pro" }, // solar
-                { family: "Ubuntu", variants: ["400", "700"] }, // united
-            ]
-        }));
-    }
-
     if (! devMode) {
         config.plugins.push(new LicenseWebpackPlugin());
-        config.plugins.push(new CreateFileWebpack({
-            path: '../web/dist',
-            fileName: 'fonts.license.txt',
-            content: 'https://fonts.google.com/attribution'
-        }));
     }
     return config;
 };